+++ lib/hanami/action/validatable.rb
@@ -9,12 +9,6 @@ module Hanami
     # @api private
     # @since 0.1.0
     module Validatable
-      # Defines the class name for anonymous params
-      #
-      # @api private
-      # @since 0.3.0
-      PARAMS_CLASS_NAME = "Params"
-
       # @api private
       # @since 0.1.0
       def self.included(base)
@@ -36,13 +30,10 @@ module Hanami
         # class. This constrains the validation to simple structure and type rules only. If you want
         # to use all the features of dry-validation contracts, use {#contract} instead.
         #
-        # The resulting contract becomes part of a dedicated params class for the action, inheriting
-        # from {Hanami::Action::Params}.
-        #
         # Instead of defining the params validation schema inline, you can alternatively provide a
-        # concrete params class, which should inherit from {Hanami::Action::Params}.
+        # concrete `Dry::Validation::Contract` subclass.
         #
-        # @param klass [Class,nil] a Hanami::Action::Params subclass
+        # @param klass [Class,nil] a Dry::Validation::Contract subclass
         # @param block [Proc] the params schema definition
         #
         # @return void
@@ -62,9 +53,6 @@ module Hanami
         #     end
         #
         #     def handle(req, *)
-        #       puts req.params.class            # => Signup::Params
-        #       puts req.params.class.superclass # => Hanami::Action::Params
-        #
         #       puts req.params[:first_name]     # => "Luca"
         #       puts req.params[:admin]          # => nil
         #     end
@@ -73,7 +61,7 @@ module Hanami
         # @example Concrete class
         #   require "hanami/action"
         #
-        #   class SignupParams < Hanami::Action::Params
+        #   class SignupContract < Dry::Validation::Contract
         #     params do
         #       required(:first_name)
         #       required(:last_name)
@@ -82,12 +70,9 @@ module Hanami
         #   end
         #
         #   class Signup < Hanami::Action
-        #     params SignupParams
+        #     params SignupContract
         #
         #     def handle(req, *)
-        #       puts req.params.class            # => SignupParams
-        #       puts req.params.class.superclass # => Hanami::Action::Params
-        #
         #       req.params[:first_name]          # => "Luca"
         #       req.params[:admin]               # => nil
         #     end
@@ -96,15 +81,7 @@ module Hanami
         # @api public
         # @since 0.3.0
         def params(klass = nil, &block)
-          contract_class =
-            if klass.nil?
-              Class.new(Dry::Validation::Contract) { params(&block) }
-            elsif klass < Params
-              # Handle subclasses of Hanami::Action::Params.
-              klass._contract.class
-            else
-              klass
-            end
+          contract_class = klass || Class.new(Dry::Validation::Contract) { params(&block) }
 
           config.contract_class = contract_class
         end
@@ -118,13 +95,10 @@ module Hanami
         # The given block is evaluated inside a `Dry::Validation::Contract` class. This allows you
         # to use all features of dry-validation contracts
         #
-        # The resulting contract becomes part of a dedicated params class for the action, inheriting
-        # from {Hanami::Action::Params}.
+        # Instead of defining the contract inline, you can alternatively provide a concrete
+        # `Dry::Validation::Contract` subclass.
         #
-        # Instead of defining the params validation contract inline, you can alternatively provide a
-        # concrete params class, which should inherit from {Hanami::Action::Params}.
-        #
-        # @param klass [Class,nil] a Hanami::Action::Params subclass
+        # @param klass [Class,nil] a Dry::Validation::Contract subclass
         # @param block [Proc] the params schema definition
         #
         # @return void
@@ -150,9 +124,6 @@ module Hanami
         #     end
         #
         #     def handle(req, *)
-        #       puts req.params.class            # => Signup::Params
-        #       puts req.params.class.superclass # => Hanami::Action::Params
-        #
         #       puts req.params[:first_name]     # => "Luca"
         #       puts req.params[:admin]          # => nil
         #     end
@@ -161,27 +132,22 @@ module Hanami
         # @example Concrete class
         #   require "hanami/action"
         #
-        #   class SignupParams < Hanami::Action::Params
-        #     contract do
-        #       params do
-        #         required(:first_name)
-        #         required(:last_name)
-        #         required(:email)
-        #       end
+        #   class SignupContract < Dry::Validation::Contract
+        #     params do
+        #       required(:first_name)
+        #       required(:last_name)
+        #       required(:email)
+        #     end
         #
-        #       rule(:email) do
-        #         # custom rule logic here
-        #       end
+        #     rule(:email) do
+        #       # custom rule logic here
         #     end
         #   end
         #
         #   class Signup < Hanami::Action
-        #     params SignupParams
+        #     contract SignupContract
         #
         #     def handle(req, *)
-        #       puts req.params.class            # => SignupParams
-        #       puts req.params.class.superclass # => Hanami::Action::Params
-        #
         #       req.params[:first_name]          # => "Luca"
         #       req.params[:admin]               # => nil
         #     end